+++ build.gradle
@@ -5,7 +5,6 @@ buildscript {
         jcenter()
     }
     dependencies {
-        classpath 'com.google.gradle:osdetector-gradle-plugin:1.4.0'
         classpath 'com.android.tools.build:gradle:2.2.2' // jcenter has the latest
         classpath 'com.github.dcendents:android-maven-gradle-plugin:1.5'
     }
