+++ common/src/jni/main/include/NativeCrypto.h
@@ -1,0 +1,41 @@
+/*
+ * Copyright (C) 2017 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef CONSCRYPT_NATIVECRYPTO_H_
+#define CONSCRYPT_NATIVECRYPTO_H_
+
+#include <jni.h>
+
+namespace conscrypt {
+
+/**
+ * Provides registration support for all native methods exported by Conscrypt.
+ */
+class NativeCrypto {
+private:
+    NativeCrypto() {}
+    ~NativeCrypto() {}
+
+public:
+    /**
+     * Register all native JNI methods exported by Conscrypt.
+     */
+    static void registerNativeMethods(JNIEnv* env);
+};
+
+}  // namespace conscrypt
+
+#endif  // CONSCRYPT_NATIVECRYPTO_H_
