+++ openjdk/build.gradle
@@ -1,23 +1,37 @@
-apply plugin: "com.google.osdetector"
-
 description = 'Conscrypt: OpenJdk'
 
 ext {
     jniSourceDir = "$rootDir/common/src/jni"
-    jniLibPrefix = "conscrypt_openjdk_jni-${osdetector.os}"
-
     assert file("$jniSourceDir").exists()
+
+    // Build the list of classifiers that will be used in the build.
+    os = org.gradle.internal.os.OperatingSystem.current();
+    if (os.isLinux()) {
+        osName = "linux"
+    } else if (os.isMacOsX()) {
+        osName = "osx"
+    } else if (os.isWindows()) {
+        osName = "windows"
+    } else {
+        throw new GradleException("Unsupported os: " + os.name)
+    }
+    nativeClassifiers = []
+    if (build32Bit) {
+        nativeClassifiers += "${osName}-x86"
+    }
+    if (build64Bit) {
+        nativeClassifiers += "${osName}-x86_64"
+    }
 }
 
 sourceCompatibility = JavaVersion.VERSION_1_8
 targetCompatibility = JavaVersion.VERSION_1_8
 
-sourceSets.main {
-    java {
-        srcDirs = [
-                "${rootDir}/common/src/main/java",
-                "src/main/java"
-        ]
+sourceSets {
+    main {
+        java {
+            srcDirs += "${rootDir}/common/src/main/java"
+        }
     }
 }
 
@@ -30,10 +44,36 @@ dependencies {
                 libraries.mockito
 }
 
-// Set the platform-specific classifier on the artifact.
-jar {
-    classifier = osdetector.classifier
+/**
+ * Create Jar and Test tasks for each native classifier.
+ */
+nativeClassifiers.each { nativeClassifier ->
+    // Normalize the classifier so that it may be used in task names. We'll also use
+    // the normalized name for the resource directory.
+    def normalizedClassifier = "${nativeClassifier}".replaceAll("-", "_")
+    def nativeFiles = files("${buildDir}/${normalizedClassifier}/resources/main")
+
+    // Create a JAR for this configuration and add it to the output archives.
+    def jarTaskName = "create${normalizedClassifier}Jar"
+    task "$jarTaskName"(type: Jar, dependsOn: classes) {
+        from sourceSets.main.output + nativeFiles
+        manifest = jar.manifest
+        classifier = nativeClassifier
+    }
+    artifacts.add('archives', tasks["$jarTaskName"])
+
+    // Create the test task and have it auto run whenever the test task runs.
+    def testTaskName = "${normalizedClassifier}Test"
+    task "$testTaskName"(type: Test) {
+        testClassesDir = test.testClassesDir
+        classpath = test.classpath + nativeFiles
+    }
+    test.dependsOn "$testTaskName"
 }
+
+// Exclude all test classes from the default test suite.
+// We will test each available native artifact separately (see nativeClassifiers).
+test.exclude("**")
 
 model {
     platforms {
@@ -47,7 +87,7 @@ model {
 
     components {
         // Builds the JNI library.
-        "$jniLibPrefix"(NativeLibrarySpec) {
+        conscrypt(NativeLibrarySpec) {
             if (build32Bit) { targetPlatform "x86" }
             if (build64Bit) { targetPlatform "x86_64" }
 
@@ -75,7 +115,7 @@ model {
                         libPath = "$boringssl64BuildDir"
                     } else {
                         throw new GradleException("Unknown architecture: " +
-                                targetPlatform.getArchitecture().getName())
+                                targetPlatform.getArchitecture().name)
                     }
 
                     if (toolChain in Clang || toolChain in Gcc) {
@@ -149,19 +189,23 @@ model {
 
     tasks {
         $.binaries.withType(SharedLibraryBinarySpec).each {
+            // Build the native artifact classifier from the OS and architecture.
             def archName = it.targetPlatform.architecture.name.replaceAll('-', '_')
+            def classifier = "${osName}-${archName}"
+            def normalizedClassifier = "$classifier".replaceAll('-', '_')
             def source = it.sharedLibraryFile
 
             // Copies the native library to a resource location that will be included in the jar.
-            task "copyNativeLib${archName}"(type: Copy) {
+            task "copyNativeLib${normalizedClassifier}"(type: Copy) {
                 from source
-                rename '(.+)(\\.[^\\.]+)', "\$1-$archName\$2"
+                // Rename the artifact to include the generated classifier
+                rename '(.+)(\\.[^\\.]+)', "\$1-$classifier\$2"
                 // This location will automatically be included in the jar.
-                into "build/resources/main/META-INF/native"
+                into "${buildDir}/${normalizedClassifier}/resources/main/META-INF/native"
             }
 
             // Make sure we build and copy the native library to the output directory.
-            compileJava.dependsOn "copyNativeLib${archName}"
+            compileJava.dependsOn "copyNativeLib${normalizedClassifier}"
         }
     }
 }
